Two Symbol Permutation Testing
In the previous version of permutation tests, the test only made pairwise comparisons of the two symbol schema and compared that against the prime implicants. This meant that for two symbol schema with three or more elements, or two different symmetrical components, that the two symbol schema could have more Prime Implicants than reality. This change checks and makes sure that the two symbol schema is contained within the prime implicants. This configuration should not be used when calculating the input symmetry, K_s, as it will produce a severe under estimate of K_s, but it should  be used when calculating the DCM, as the DCM should be a faithful representation of the logic gate. As such, I've added in an argument that gets passed through starting with the node's canalizing map function, to the two symmetry calculation. I did not add handling to check if the two symbol was produced under which algorithm (for example, someone could call check compute two symbol and then call to produce a DCM and the result would be different than just calling DCM, similarly saved data does not know which algorithm was used).