change the behavior for thresholded effective graph
Previously, no matter thresholded or not, the effective graph is always cached.
If the user call effective_graph() repeatedly with different threshold, actually
they will get the same effective graph. This behavior is very confusing for the
user, since most of them will not expect to purge self._eg or to manually make a
copy of self._eg.

The new behavior: the vanilla effective graph will be cached and calculated only
once. On the contrary, all thresholded effective graph will be copied from the
vanilla effective graph and return to the user.